+++ src/ShapeCrawler/Presentations/SPresentationPart.cs
@@ -16,13 +16,13 @@ internal readonly ref struct SPresentationPart
 
     internal void AddSlidePart(SlidePart slidePart)
     {
-        var rId = this.presentationPart.NextRelationshipId();
+        var rId = new SOpenXmlPart(this.presentationPart).NextRelationshipId();
         var addedSlidePart = this.presentationPart.AddPart(slidePart, rId);
 
         var notesSlidePartAddedSlidePart = addedSlidePart.GetPartsOfType<NotesSlidePart>().FirstOrDefault();
         notesSlidePartAddedSlidePart?.DeletePart(notesSlidePartAddedSlidePart.NotesMasterPart!);
 
-        rId = this.presentationPart.NextRelationshipId();
+        rId = new SOpenXmlPart(this.presentationPart).NextRelationshipId();
         var addedSlideMasterPart = this.presentationPart.AddPart(addedSlidePart.SlideLayoutPart!.SlideMasterPart!, rId);
         var layoutIdList = addedSlideMasterPart.SlideMaster.SlideLayoutIdList!.OfType<P.SlideLayoutId>();
         foreach (var layoutId in layoutIdList.ToList())
