+++ src/ShapeCrawler/Presentations/Slide.cs
@@ -226,7 +226,7 @@ internal sealed class Slide : ISlide
         }
 
         // https://learn.microsoft.com/en-us/office/open-xml/presentation/working-with-notes-slides
-        var rid = this.SdkSlidePart.NextRelationshipId();
+        var rid = new SOpenXmlPart(this.SdkSlidePart).NextRelationshipId();
         var notesSlidePart1 = this.SdkSlidePart.AddNewPart<NotesSlidePart>(rid);
         var notesSlide = new NotesSlide(
             new CommonSlideData(
