+++ src/ShapeCrawler/Presentations/Slides.cs
@@ -73,7 +73,7 @@ internal sealed class Slides : ISlides
     {
         var sdkPresDocument = (PresentationDocument)this.presentationPart.OpenXmlPackage;
         var sdkPresPart = sdkPresDocument.PresentationPart!;
-        var rId = sdkPresPart.NextRelationshipId();
+        var rId = new SOpenXmlPart(sdkPresPart).NextRelationshipId();
         var sdkSlidePart = sdkPresPart.AddNewPart<SlidePart>(rId);
         sdkSlidePart.Slide = new P.Slide(
             new P.CommonSlideData(
@@ -86,8 +86,9 @@ internal sealed class Slides : ISlides
             new P.ColorMapOverride(new A.MasterColorMapping()));
         var layoutInternal = (SlideLayout)layout;
         sdkSlidePart.AddPart(layoutInternal.SdkSlideLayoutPart(), "rId1");
-        
-        if (layoutInternal.SdkSlideLayoutPart().SlideLayout.CommonSlideData is P.CommonSlideData commonSlideData && commonSlideData.ShapeTree is P.ShapeTree shapeTree) 
+
+        if (layoutInternal.SdkSlideLayoutPart().SlideLayout.CommonSlideData is P.CommonSlideData commonSlideData &&
+            commonSlideData.ShapeTree is P.ShapeTree shapeTree)
         {
             var placeholderShapes = shapeTree.ChildElements
                 .OfType<P.Shape>()
@@ -153,7 +154,8 @@ internal sealed class Slides : ISlides
         var sourceSlideId = (P.SlideId)sourceSlidePresPart.Presentation.SlideIdList!.ChildElements[slide.Number - 1];
         var sourceSlidePart = (SlidePart)sourceSlidePresPart.GetPartById(sourceSlideId.RelationshipId!);
 
-        new SSlideMasterPart(sourceSlidePart.SlideLayoutPart!.SlideMasterPart!).RemoveLayoutsExcept(sourceSlidePart.SlideLayoutPart!);
+        new SSlideMasterPart(sourceSlidePart.SlideLayoutPart!.SlideMasterPart!).RemoveLayoutsExcept(sourceSlidePart
+            .SlideLayoutPart!);
 
         var wrappedPresentationPart = new SPresentationPart(targetPresPart);
         wrappedPresentationPart.AddSlidePart(sourceSlidePart);
@@ -170,7 +172,9 @@ internal sealed class Slides : ISlides
         // Creates a new TextBody
         if (shape.TextBody is null)
         {
-            return new P.TextBody(new DocumentFormat.OpenXml.Drawing.Paragraph([new DocumentFormat.OpenXml.Drawing.EndParagraphRunProperties()]))
+            return new P.TextBody(new DocumentFormat.OpenXml.Drawing.Paragraph([
+                new DocumentFormat.OpenXml.Drawing.EndParagraphRunProperties()
+            ]))
             {
                 BodyProperties = new DocumentFormat.OpenXml.Drawing.BodyProperties(),
                 ListStyle = new DocumentFormat.OpenXml.Drawing.ListStyle(),
@@ -179,7 +183,7 @@ internal sealed class Slides : ISlides
 
         return (P.TextBody)shape.TextBody.CloneNode(true);
     }
-    
+
     private static void AdjustLayoutIds(PresentationDocument sdkPresDocDest, uint masterId)
     {
         foreach (var slideMasterPart in sdkPresDocDest.PresentationPart!.SlideMasterParts)
@@ -203,8 +207,7 @@ internal sealed class Slides : ISlides
         var masterId = CreateId(sdkPresDest.SlideMasterIdList!);
         P.SlideMasterId slideMaterId = new()
         {
-            Id = masterId,
-            RelationshipId = sdkPresDocDest.PresentationPart!.GetIdOfPart(addedSlideMasterPart!)
+            Id = masterId, RelationshipId = sdkPresDocDest.PresentationPart!.GetIdOfPart(addedSlideMasterPart!)
         };
         sdkPresDocDest.PresentationPart.Presentation.SlideMasterIdList!.Append(slideMaterId);
         sdkPresDocDest.PresentationPart.Presentation.Save();
@@ -234,7 +237,7 @@ internal sealed class Slides : ISlides
 
         return ++currentId;
     }
-    
+
     private static uint CreateId(P.SlideMasterIdList slideMasterIdList)
     {
         uint currentId = 0;
